requirements.txt
bcrypt==4.2.0
beautifulsoup4==4.13.0b2
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssmin==0.2.0
dnspython==2.6.1
eventlet==0.37.0
Flask==3.0.3
Flask-Assets==2.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
mailersend==0.5.8
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.0
webassets==2.0
Werkzeug==3.0.4
